Create README.md
This thesis aims to analyze the commitments of nations in the Glasgow Pact concerning the energy sector. The study takes a critical look at the limits of this type of international governance, whose negotiations are long, subject to power struggles and ideological biases, thus weakening its real impact. We distinguish two levers of action that were plebiscited at the 26th Conference of the Parties to reduce the carbon footprint of energy consumption:
technology and "green" finance. These commitments are largely influenced by neoliberal, capitalist, and techno-solutionist ideologies. The objective of these two levers would be to limit the damage of a society at the service of the growth of capital without questioning our development choices. The regulation of markets or measures of sobriety are absent, alternative levers that we will rehabilitate in this article. In order to fuel a debate that was absent at the
26th Conference of the Parties, this work proposes to take a critical look at the proposed levers and their theoretical and practical limits.